manus-backend/requirements.txt
blinker==1.9.0
click==8.2.1
Flask==3.1.1
flask-cors==6.0.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.6.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
SQLAlchemy==2.0.41
typing_extensions==4.14.0
Werkzeug==3.1.3

# Environment and configuration
python-dotenv==1.0.0

# Authentication and security
bcrypt==4.1.2
PyJWT==2.8.0

# AI and ML APIs
openai==1.12.0
anthropic==0.18.1
google-generativeai==0.4.0

# HTTP requests and API clients
requests==2.31.0
httpx==0.27.0

# Task queue and background processing
celery==5.3.6
redis==5.0.1

# Database drivers
psycopg2-binary==2.9.9
pymongo==4.6.1

# Data processing and validation
pydantic==2.6.1
marshmallow==3.20.2

# File handling and storage
Pillow==10.2.0
boto3==1.34.34

# Web scraping and automation
beautifulsoup4==4.12.3
selenium==4.17.2

# Utilities
python-dateutil==2.8.2
pytz==2024.1
uuid==1.30
